+++ sdk/eventhub/azure-eventhub-checkpointstoreblob-aio/samples/event_processor_blob_storage_example_with_storage_api_version.py
@@ -6,6 +6,7 @@ from azure.eventhub.extensions.checkpointstoreblobaio import BlobCheckpointStore
 CONNECTION_STR = os.environ["EVENT_HUB_CONN_STR"]
 STORAGE_CONNECTION_STR = os.environ["AZURE_STORAGE_CONN_STR"]
 BLOB_CONTAINER_NAME = "your-blob-container-name"  # Please make sure the blob container resource exists.
+STORAGE_SERVICE_API_VERSION = "2017-11-09"
 
 
 async def on_event(partition_context, event):
@@ -24,6 +25,7 @@ if __name__ == '__main__':
     checkpoint_store = BlobCheckpointStore.from_connection_string(
         STORAGE_CONNECTION_STR,
         container_name=BLOB_CONTAINER_NAME,
+        api_version=STORAGE_SERVICE_API_VERSION
     )
     client = EventHubConsumerClient.from_connection_string(
         CONNECTION_STR,