+++ sdk/eventhub/azure-eventhub-checkpointstoreblob/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.0.1 (Unreleased)
 
 **New features**
-- Param `api_version` of `BlobPartitionManager` now supports older versions of Azure Storage Serice API.
+- Param `api_version` of `BlobPartitionManager` now supports older versions of Azure Storage Service API.
 
 ## 1.0.0 (2020-01-13)
 Stable release. No new features or API changes.